Performance Tab improvement
Converted to a system report.  This provides some benefits like being able to expand charts. Report is now driven by metadata instead of code.

+++ DBADashGUI/Performance/DrivePerformance.cs
@@ -19,7 +19,7 @@ namespace DBADashGUI.Performance
         private int mins;
         private int dateGrouping;
         private double LatencyLimit = 200;
-        private int PointSize => dataPointsToolStripMenuItem.Checked ? 10 : 0;
+        private double PointSize => dataPointsToolStripMenuItem.Checked ? 10 : 0;
 
         private void TsDateGroup_Click(object sender, EventArgs e)
         {
